server/requirements.txt
aiohttp
python-socketio
orjson
setproctitle